requirements.txt
pyopengl
pygame
common-game-maths
Pillow